requirements.txt
qdrant-client
openai
python-dotenv
tqdm
pydantic
os
embeddings